I wouldn't release this if I were you.
This is mainly a migration of Computronics to OpenComputers 1.3, to get
it ready for when the mod goes out of beta. The APIs changed a little,
so I had to change the usage of Microchips in crafting recipes, and a
bunch of stuff in the Camera upgrade (which is a lot of casting, since
classes and methods were changed). These changes SHOULD be migrated to
the 1.6.4 branch, but that branch is set up differently. Unless you want
to gradleize it, that is.
I also fixed your fix of my derp from earlier. Now, the new tape will
show up in the Creative menu, even if GregTech isn't present.
And there's some IntelliJ stuff in here now, since I'm switching over to
that IDE instead of just manually hacking away in Notepad++ :V