Not releasable yet, again
Same deal as in the 1.7.2 branch, this is stuff for supporting OC1.3,
which isn't out yet. Also, it's untested, since I don't have your build
environment (but gamax92 should be making a PR soon with that)
The API you need to build this with is here:
http://ci.cil.li/job/OpenComputers/ anything from the 1.3.0 releases
should do the trick.
Also, brought over the new Nether Star tape from the 1.7.2 branch!