Restored Creative Chat Box... sort of.
I herped and I derped and I forgot to bring this block over when porting
the update. HOWEVER for some reason I can't get it to show up in NEI or
the Creative mode inventory, but you CAN /give it to yourself.
Uncommenting line 24 in ItemBlockChatBox.java will cause compile to
fail... but this is probably where the problem is.