Radar block!!!
- Yep, radar block implemented. Its primary (and only) function is
getEntities(), which seeks out all living entities within a configurable
range (default is 32 blocks in X and Z directions) and return the
results in a table showing the name and coordinates of the entity.
- For some reason, it's not picking up players? Players must not be
considered living entities then :V Will be looking into this one.
- Tweaked jaquadro's fix from last night to actually compile properly.
But it seems that the pick block action isn't getting the subBlock?