Attempted to fix filesystem, see description
<@Sangar> mm. something to try: in initOCFilesystem, set the env's
node's visibility to network (defaults to neighbors) and remove the
manual connects in the onConnect and onDisconnect
<@Sangar> (those
connects are really evil btw, because that directly connects it to
*every node* in the network like that, which will lead to magic
connections even if a cable is broken :P))

<@Sangar> onConnect is
called for *every* reachable node in the network. so if you have a
network A-B-C-TapeDrive, the node gets directly connected to A,B and C.
No remove C. Normally that'd split the tape drive from A and B. but
because they're now directly connected, that split won't
happen.

<@Sangar> if anything you can do something like check if the
passed node is a neighbor of the env, and if yes connect it to the fs
node
<@Sangar> that would be safe

<@Sangar> but that's just another
issue anyway, i don't know why the fs won't show up. with that much
overkill it certainly should :/
<@Sangar> could you set a breakpoint in
the initOCFilesystem and check if oc_fs is initialized to something...
sane?
<Techokami> Sangar here's the minecraft log when I placed down a
Tape Drive http://pastebin.com/AW6p5HND
<@Sangar> well, it's telling you
what the problem is :P
<Techokami> so how do I make it not connect to
itself?
<@Sangar> well, the simplest fix is `if (node != oc_fs.node())
node.connect(oc_fs.node())`. but as i said, you'll either want to make
it network visible and remove that entirely, or limit it to connect it
to neighbors only, anyway, both of which will solve this :).