Yet more attempts at fixing filesystem
Sangar and I tried our damndest to suss out the issue, but we could not
get the damn filesystem thing to work. And I am tired. Gruh.
I moved tape.lua to a different path. I know it's not the same path as
you had before, but it's the path used by another mod that was able to
successfully use the filesystem, so I had to try.