OpenComputers 1.4 support
This is a quick implementation of the current (alpha) state of the new
OC 1.4 APIs, which would make this an alpha-quality codebase of
Computronics. Provided for the super early adopters to get their fix
with the alpha release, and to be prepared for OC 1.4 when it goes into
beta, then stable release.
BIG THANKS TO VEXATOS! You helped me out man.