Added code for saving and loading instruction queues
Added a method for storing an instruction in NBT.
Added a method for storing the instruction queue in NBT.
Cleaned up the code for loading from NBT, there is no need for an
additional CompoundTag, just reuse the main one instead.
Made SetADSR's NBT constructor reuse the main constructor to properly
store the input settings for the ADSR.

Demote channel index variable type to a byte.

Fixed a small bug in AudioUtil, the creation of an AudioProcess would
ignore the number of channels specified to it.